+++ src/main/java/com/comandante/creeper/player/Levels.java
@@ -7,14 +7,14 @@ public class Levels {
 
     private static double CONSTANT_MODIFIER = 0.005;
 
-    public static int getLevel(int experience) {
+    public static long getLevel(long experience) {
         double v = CONSTANT_MODIFIER * sqrt(experience);
-        return (int) Math.floor(v);
+        return (long) Math.floor(v);
     }
 
-    public static int getXp(int level) {
+    public static long getXp(long level) {
         double v = pow(level, 2) / pow(CONSTANT_MODIFIER, 2);
-        return (int) Math.ceil(v);
+        return (long) Math.ceil(v);
     }
 
 }
